+++ lib/src/mutations.dart
@@ -62,8 +62,10 @@ class MutationIterator implements Iterator<MutatedCode> {
   int _index = 0;
   bool _initialized = false;
 
-  final MutatedCode _currentMutation =
-      MutatedCode('', MutatedLine(0, 0, 0, '', ''));
+  final MutatedCode _currentMutation = MutatedCode(
+    '',
+    MutatedLine(0, 0, 0, '', '', Mutation('')),
+  );
   final Iterator<Match> _matches;
 
   @override
@@ -87,12 +89,15 @@ class MutationIterator implements Iterator<MutatedCode> {
     }
     _currentMutation.text =
         mutation.replacements[_index].replace(text, _matches.current);
+
     _currentMutation.line = createMutatedLine(
       _matches.current.start,
       _matches.current.end,
       text,
       _currentMutation.text,
+      mutation,
     );
+
     _index += 1;
     return true;
   }
@@ -129,6 +134,7 @@ MutatedLine createMutatedLine(
   int absoluteEnd,
   String original,
   String mutated,
+  Mutation mutation,
 ) {
   if (absoluteStart < 0) absoluteStart = 0;
   if (absoluteStart > original.length) absoluteStart = original.length;
@@ -157,12 +163,20 @@ MutatedLine createMutatedLine(
     mutationEnd,
     original.substring(lineStart, lineEnd),
     mutated.substring(lineStart, lineEndMutated),
+    mutation,
   );
 }
 
 /// A mutation data structure with Information about a mutated line.
 class MutatedLine {
-  MutatedLine(this.line, int first, int last, this.original, this.mutated) {
+  MutatedLine(
+    this.line,
+    int first,
+    int last,
+    this.original,
+    this.mutated,
+    this.mutation,
+) {
     /// make wrong states impossible to repesent
     start = first >= 0 ? first : 0;
     end = last <= original.length ? last : original.length;
@@ -182,6 +196,8 @@ class MutatedLine {
 
   /// mutated line of code
   final String mutated;
+
+  final Mutation mutation;
 
   /// Pretty formatting
   String toMarkdown() {